requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg[binary]==3.2.2
itsdangerous>=2.1
bcrypt==4.1.3
